remove: 게시글 Soft Deletion 구현
- 타인의 댓글이 존재할 때 게시글을 삭제할 수 없도록 구현했습니다.
- 자신의 댓글만 존재하면 게시글을 삭제할 수 있습니다.
- 이에 따른 테스트를 추가했습니다.

+++ src/main/java/codesquad/jspcafe/domain/article/service/ArticleService.java
@@ -5,6 +5,8 @@ import codesquad.jspcafe.domain.article.payload.request.ArticleUpdateRequest;
 import codesquad.jspcafe.domain.article.payload.response.ArticleCommonResponse;
 import codesquad.jspcafe.domain.article.payload.response.ArticleContentResponse;
 import codesquad.jspcafe.domain.article.repository.ArticleRepository;
+import codesquad.jspcafe.domain.reply.domain.Reply;
+import codesquad.jspcafe.domain.reply.repository.ReplyRepository;
 import codesquad.jspcafe.domain.user.domain.User;
 import codesquad.jspcafe.domain.user.repository.UserRepository;
 import java.nio.file.AccessDeniedException;
@@ -17,10 +19,13 @@ import java.util.Objects;
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final ReplyRepository replyRepository;
     private final UserRepository userRepository;
 
-    public ArticleService(ArticleRepository articleRepository, UserRepository userRepository) {
+    public ArticleService(ArticleRepository articleRepository, ReplyRepository replyRepository,
+        UserRepository userRepository) {
         this.articleRepository = articleRepository;
+        this.replyRepository = replyRepository;
         this.userRepository = userRepository;
     }
 
@@ -57,6 +62,11 @@ public class ArticleService {
         if (!Objects.equals(article.getWriter().getId(), userId)) {
             throw new AccessDeniedException("삭제 권한이 없습니다.");
         }
+        List<Reply> replies = replyRepository.findByArticleId(articleId);
+        if (!replies.stream().allMatch(reply -> reply.getWriter().getId().equals(userId))) {
+            throw new IllegalArgumentException("댓글이 존재하여 삭제할 수 없습니다.");
+        }
+        replies.forEach(replyRepository::delete);
         articleRepository.delete(article);
     }
